Revamped host categorization and moved to Google Safe Browsing for detecting safe and malware sites

+++ src/Host.cpp
@@ -61,18 +61,18 @@ Host::~Host() {
 
   if(ip && (!ip->isEmpty())) dumpStats(false);
 
-  k = get_string_key(host_key, sizeof(host_key));
-
-  /* Host Contacts */
-  if(k && (k[0] != '\0')) {
-    snprintf(key, sizeof(key), "%s.client", k);
-    ntop->getRedis()->del(key);
-
-    snprintf(key, sizeof(key), "%s.server", k);
-    ntop->getRedis()->del(key);
-  }
-
   if(localHost || systemHost) {
+    k = get_string_key(host_key, sizeof(host_key));
+
+    /* Host Contacts */
+    if(k && (k[0] != '\0')) {
+      snprintf(key, sizeof(key), "%s.client", k);
+      ntop->getRedis()->del(key);
+
+      snprintf(key, sizeof(key), "%s.server", k);
+      ntop->getRedis()->del(key);
+    }
+
     if(ip != NULL) {
       snprintf(key, sizeof(key), "%s.client", k);
       ntop->getRedis()->del(key);
@@ -80,9 +80,9 @@ Host::~Host() {
       snprintf(key, sizeof(key), "%s.server", k);
       ntop->getRedis()->del(key);
     }
-  }
 
-  dumpHostContacts(HOST_FAMILY_ID);
+    dumpHostContacts(HOST_FAMILY_ID);
+  }
 
   // ntop->getTrace()->traceEvent(TRACE_NORMAL, "Deleting %s (%s)", k, localHost ? "local": "remote");
 
@@ -140,15 +140,17 @@ void Host::computeHostSerial() {
 /* *************************************** */
 
 void Host::flushContacts(bool freeHost) {
-  dumpHostContacts(HOST_FAMILY_ID);
-  contacts->purgeAll();
+  if(localHost || systemHost) {
+    dumpHostContacts(HOST_FAMILY_ID);
+    contacts->purgeAll();
 
-  if(!freeHost) {
-    /*
-      Recompute it so that if the day wrapped
-      we have a new one
-    */
-    computeHostSerial();
+    if(!freeHost) {
+      /*
+	Recompute it so that if the day wrapped
+	we have a new one
+      */
+      computeHostSerial();
+    }
   }
 }
 
@@ -308,23 +310,25 @@ void Host::updateHostTrafficPolicy(char *key) {
 
 void Host::updateHostL7Policy() {
 #ifdef NTOPNG_PRO
+  if(!iface->is_bridge_interface()) return;
+
   if(ntop->getPro()->has_valid_license()) {
-    if(localHost && (ip != NULL)) {
+    if((localHost || systemHost) && ip) {
       char host_key[64], hash_name[64], rsp[32];
       char buf[64], *host = ip->print(buf, sizeof(buf));
 
       l7Policy = getInterface()->getL7Policer()->getIpPolicy(ip, vlan_id);
-      
-      snprintf(host_key, sizeof(host_key), "%s/%u@%u", host, 
+
+      snprintf(host_key, sizeof(host_key), "%s/%u@%u", host,
 	       ip->isIPv4() ? 32 : 128, vlan_id);
 
       /* ************************************************* */
 
-      snprintf(hash_name, sizeof(hash_name), 
+      snprintf(hash_name, sizeof(hash_name),
 	       "ntopng.prefs.%u.l7_policy_ingress_shaper_id",
 	       getInterface()->get_id());
 
-      if((ntop->getRedis()->hashGet(hash_name, host_key, rsp, sizeof(rsp)) != 0) 
+      if((ntop->getRedis()->hashGet(hash_name, host_key, rsp, sizeof(rsp)) != 0)
 	 || (rsp[0] == '\0'))
 	ingress_shaper_id = DEFAULT_SHAPER_ID;
       else {
@@ -336,11 +340,11 @@ void Host::updateHostL7Policy() {
 
       /* ************************************************* */
 
-      snprintf(hash_name, sizeof(hash_name), 
+      snprintf(hash_name, sizeof(hash_name),
 	       "ntopng.prefs.%u.l7_policy_egress_shaper_id",
 	       getInterface()->get_id());
 
-      if((ntop->getRedis()->hashGet(hash_name, host_key, rsp, sizeof(rsp)) != 0) 
+      if((ntop->getRedis()->hashGet(hash_name, host_key, rsp, sizeof(rsp)) != 0)
 	 || (rsp[0] == '\0'))
 	egress_shaper_id = DEFAULT_SHAPER_ID;
       else {
@@ -404,8 +408,8 @@ void Host::set_mac(char *m) {
          &mac[0], &mac[1], &mac[2], &mac[3], &mac[4], &mac[5]);
 
   mac_address[0] = mac[0], mac_address[1] = mac[1],
-	  mac_address[2] = mac[2], mac_address[3] = mac[3],
-	  mac_address[4] = mac[4], mac_address[5] = mac[5];
+    mac_address[2] = mac[2], mac_address[3] = mac[3],
+    mac_address[4] = mac[4], mac_address[5] = mac[5];
 }
 
 /* *************************************** */
@@ -430,7 +434,7 @@ void Host::lua(lua_State* vm, patricia_tree_t *ptree,
 
   lua_push_int_table_entry(vm, "bytes.sent", sent.getNumBytes());
   lua_push_int_table_entry(vm, "bytes.rcvd", rcvd.getNumBytes());
-  
+
   if(ip) {
     char *local_net;
 
@@ -545,7 +549,7 @@ void Host::lua(lua_State* vm, patricia_tree_t *ptree,
       char *rsp = serialize();
 
       if(ndpiStats) ndpiStats->lua(iface->get_view(), vm);
-      getHostContacts(vm, ptree);
+      if(localHost || systemHost) getHostContacts(vm, ptree);
       lua_push_str_table_entry(vm, "json", rsp);
       free(rsp);
 
@@ -560,16 +564,16 @@ void Host::lua(lua_State* vm, patricia_tree_t *ptree,
       if(ntop->getPro()->has_valid_license()) {
 	if(l7Policy != NULL) {
 	  lua_newtable(vm);
-	  
+
 	  for(int i=1; i<NDPI_MAX_SUPPORTED_PROTOCOLS+NDPI_MAX_NUM_CUSTOM_PROTOCOLS; i++) {
 	    if(NDPI_ISSET(l7Policy, i)) {
 	      char *proto = ndpi_get_proto_by_id(iface->get_ndpi_struct(), i);
-	      
+
 	      if(proto)
 		lua_push_int_table_entry(vm, proto, i);
 	    }
 	  }
-	  
+
 	  lua_pushstring(vm, "l7_traffic_policy");
 	  lua_insert(vm, -2);
 	  lua_settable(vm, -3);
@@ -603,7 +607,7 @@ void Host::lua(lua_State* vm, patricia_tree_t *ptree,
     lua_settable(vm, -3);
   }
 
-  
+
 
 }
 
@@ -730,9 +734,11 @@ u_int32_t Host::key() {
 /* *************************************** */
 
 void Host::incrContact(Host *_peer, bool contacted_peer_as_client) {
-  if(_peer->get_ip() != NULL)
-    ((GenericHost*)this)->incrContact(iface, get_host_serial(), _peer->get_ip(),
-				      contacted_peer_as_client);
+  if(localHost || systemHost) {
+    if(_peer->get_ip() != NULL)
+      ((GenericHost*)this)->incrContact(iface, get_host_serial(), _peer->get_ip(),
+					contacted_peer_as_client);
+  }
 }
 
 /* *************************************** */
@@ -826,7 +832,7 @@ char* Host::serialize() {
   json_object_object_add(my_object, "sent", sent.getJSONObject());
   json_object_object_add(my_object, "rcvd", rcvd.getJSONObject());
   json_object_object_add(my_object, "ndpiStats", ndpiStats->getJSONObject(iface));
-  json_object_object_add(my_object, "contacts", contacts->getJSONObject());
+  if(localHost || systemHost) json_object_object_add(my_object, "contacts", contacts->getJSONObject());
   json_object_object_add(my_object, "activityStats", activityStats.getJSONObject());
 
   if(dns)  json_object_object_add(my_object, "dns", dns->getJSONObject());
@@ -936,7 +942,9 @@ bool Host::deserialize(char *json_str) {
   if(json_object_object_get_ex(o, "activityStats", &obj)) activityStats.deserialize(obj);
 
   computeHostSerial();
-  if(json_object_object_get_ex(o, "contacts", &obj)) contacts->deserialize(iface, this, obj);
+  if(localHost || systemHost) {
+    if(json_object_object_get_ex(o, "contacts", &obj)) contacts->deserialize(iface, this, obj);
+  }
   if(json_object_object_get_ex(o, "pktStats.sent", &obj)) sent_stats.deserialize(obj);
   if(json_object_object_get_ex(o, "pktStats.recv", &obj)) recv_stats.deserialize(obj);
 
@@ -946,11 +954,11 @@ bool Host::deserialize(char *json_str) {
   last_update_time.tv_sec = (long)time(NULL), last_update_time.tv_usec = 0;
   // Update bps throughput
   bytes_thpt = 0, last_bytes = sent.getNumBytes()+rcvd.getNumBytes(),
-  bytes_thpt_trend = trend_unknown;
+    bytes_thpt_trend = trend_unknown;
 
   // Update pps throughput
   pkts_thpt = 0, last_packets = sent.getNumPkts()+rcvd.getNumPkts(),
-  pkts_thpt_trend = trend_unknown;
+    pkts_thpt_trend = trend_unknown;
 
   return(true);
 }
@@ -968,7 +976,7 @@ void Host::updateSynFlags(time_t when, u_int8_t flags, Flow *f, bool syn_sent) {
 
 #if 0
     /*
-       It's normal that at startup several flows are created
+      It's normal that at startup several flows are created
     */
     if(ntop->getUptime() < 10 /* sec */) return;
 #endif
@@ -978,11 +986,11 @@ void Host::updateSynFlags(time_t when, u_int8_t flags, Flow *f, bool syn_sent) {
     if(syn_sent) {
       error_msg = "Host <A HREF=%s/lua/host_details.lua?host=%s&ifname=%s>%s</A> is a SYN flooder [%u SYNs sent in the last %u sec] %s";
       snprintf(msg, sizeof(msg),
-	     error_msg, ntop->getPrefs()->get_http_prefix(),
-	     h, iface->get_name(), h,
-	     counter->getCurrentHits(),
-	     counter->getOverThresholdDuration(),
-	     f->print(flow_buf, sizeof(flow_buf)));
+	       error_msg, ntop->getPrefs()->get_http_prefix(),
+	       h, iface->get_name(), h,
+	       counter->getCurrentHits(),
+	       counter->getOverThresholdDuration(),
+	       f->print(flow_buf, sizeof(flow_buf)));
     } else {
       Host *attacker = f->get_srv_host();
       IpAddress *aip = attacker->get_ip();
@@ -990,11 +998,11 @@ void Host::updateSynFlags(time_t when, u_int8_t flags, Flow *f, bool syn_sent) {
       aip_ptr = aip->print(aip_buf, sizeof(aip_buf));
       error_msg = "Host <A HREF=%s/lua/host_details.lua?host=%s&ifname=%s>%s</A> is under SYN flood attack by host %s [%u SYNs received in the last %u sec] %s";
       snprintf(msg, sizeof(msg),
-	     error_msg, ntop->getPrefs()->get_http_prefix(),
-	     h, iface->get_name(), h, aip_ptr,
-	     counter->getCurrentHits(),
-	     counter->getOverThresholdDuration(),
-	     f->print(flow_buf, sizeof(flow_buf)));
+	       error_msg, ntop->getPrefs()->get_http_prefix(),
+	       h, iface->get_name(), h, aip_ptr,
+	       counter->getCurrentHits(),
+	       counter->getOverThresholdDuration(),
+	       f->print(flow_buf, sizeof(flow_buf)));
     }
 
     ntop->getTrace()->traceEvent(TRACE_INFO, "SYN Flood: %s", msg);
@@ -1057,13 +1065,13 @@ void Host::decNumFlows(bool as_client) {
       if(num_active_flows_as_client == max_num_active_flows) {
 	const char* error_msg = "Host <A HREF=%s/lua/host_details.lua?host=%s&ifname=%s>%s</A> is no longer a possible scanner [%u active flows]";
 	char ip_buf[48], *h, msg[512];
-	
+
 	h = ip->print(ip_buf, sizeof(ip_buf));
-	
+
 	snprintf(msg, sizeof(msg),
 		 error_msg, ntop->getPrefs()->get_http_prefix(),
 		 h, iface->get_name(), h, num_active_flows_as_client);
-	
+
 	ntop->getTrace()->traceEvent(TRACE_INFO, "End scan attack: %s", msg);
 	ntop->getRedis()->queueAlert(alert_level_info, alert_flow_flood, msg);
 	incNumAlerts();
@@ -1077,13 +1085,13 @@ void Host::decNumFlows(bool as_client) {
       if(num_active_flows_as_server == max_num_active_flows) {
 	const char* error_msg = "Host <A HREF=%s/lua/host_details.lua?host=%s&ifname=%s>%s</A> is no longer under scan attack [%u active flows]";
 	char ip_buf[48], *h, msg[512];
-	
+
 	h = ip->print(ip_buf, sizeof(ip_buf));
-	
+
 	snprintf(msg, sizeof(msg),
 		 error_msg, ntop->getPrefs()->get_http_prefix(),
 		 h, iface->get_name(), h, num_active_flows_as_server);
-	
+
 	ntop->getTrace()->traceEvent(TRACE_INFO, "End scan attack: %s", msg);
 	ntop->getRedis()->queueAlert(alert_level_info, alert_flow_flood, msg);
 	incNumAlerts();
@@ -1111,7 +1119,7 @@ void Host::setQuota(u_int32_t new_quota) {
 
 bool Host::isAboveQuota() {
   return host_quota_mb > 0 /* 0 == unlimited */ &&
-         ((GenericHost*)this)->getPeriodicStats() > (host_quota_mb * 1000000);
+    ((GenericHost*)this)->getPeriodicStats() > (host_quota_mb * 1000000);
 }
 
 void Host::updateStats(struct timeval *tv) {
@@ -1196,7 +1204,7 @@ void Host::resetPeriodicStats() {
 
 /* *************************************** */
 
-void Host::updateHTTPHostRequest(char *virtual_host_name, u_int32_t num_req, 
+void Host::updateHTTPHostRequest(char *virtual_host_name, u_int32_t num_req,
 				 u_int32_t bytes_sent, u_int32_t bytes_rcvd) {
   if(http)
     http->updateHTTPHostRequest(virtual_host_name, num_req, bytes_sent, bytes_rcvd);
@@ -1223,18 +1231,20 @@ void Host::setDumpTrafficPolicy(bool new_policy) {
 
 void Host::readAlertPrefs() {
   trigger_host_alerts = false;
-  
+
   if(!localHost) return;
 
-  if(!ntop->getPrefs()->are_alerts_disabled()) {
-    char *key, ip_buf[48], rsp[32];
-    
-    key = get_string_key(ip_buf, sizeof(ip_buf));
-    if(key) {
-      ntop->getRedis()->hashGet((char*)CONST_ALERT_PREFS, key, rsp, sizeof(rsp));
+  if(ip && (!ip->isEmpty())) {
+    if(!ntop->getPrefs()->are_alerts_disabled()) {
+      char *key, ip_buf[48], rsp[32];
       
-      trigger_host_alerts = ((strcmp(rsp, "false") == 0) ? 0 : 1);
-    } else
-      trigger_host_alerts = false;
-  }  
+      key = get_string_key(ip_buf, sizeof(ip_buf));
+      if(key) {
+	ntop->getRedis()->hashGet((char*)CONST_ALERT_PREFS, key, rsp, sizeof(rsp));
+	
+	trigger_host_alerts = ((strcmp(rsp, "false") == 0) ? 0 : 1);
+      } else
+	trigger_host_alerts = false;
+    }
+  }
 }
